+++ docs/responses.md
@@ -1,0 +1,8 @@
+## responses
+
+This project is (at least in part) an experiment in using the **new** respopnses API from OpenAI. 
+Note that one of the oddities of the responses API is that there are only 3 roles:
+        1. "user" → what the human types
+        2. "assistant" → what the model has said in past turns
+        3. "system" → the instructions that guide the assistant’s behavior
+Here is [the reference manual](https://platform.openai.com/docs/api-reference/responses)
